updated for sync view with color images
